train_model function, move mambamodel class to own file
